requirements.txt
Flask==1.0.*

# Application server for both development and production.
gunicorn==19.8.*

# Testing and static analysis.
pytest==3.5.*
pytest-cov==2.5.*
flake8==3.5.*
pytz==2018.5

# Database
Flask-SQLAlchemy==2.3.*
pymysql==0.7.11

# Forms.
Flask-WTF==0.14.*
WTForms-Components==0.9.*

# Extensions.
flask-debugtoolbar==0.10.0
flask-less==0.4
Flask-Login==0.4.*
Flask-Limiter==1.0.*


